Moved library-handling code from legacy folder to main.c; added a check to avoid processing new protocol CLA when not in BTC or tBTC
