+++ src/legacy/main_old.c
@@ -40,9 +40,6 @@
 
 #include "../swap/swap_lib_calls.h"
 #include "../swap/btchip_bcd.h"
-#include "../swap/handle_swap_sign_transaction.h"
-#include "../swap/handle_get_printable_amount.h"
-#include "../swap/handle_check_address.h"
 
 #define __NAME3(a, b, c) a##b##c
 #define NAME3(a, b, c) __NAME3(a, b, c)
@@ -1152,59 +1149,3 @@ void btchip_bagl_request_segwit_input_approval()
     ux_flow_init(0, ux_request_segwit_input_approval_flow, NULL);
 }
 
-
-static void library_main_helper(struct libargs_s *args) {
-    check_api_level(CX_COMPAT_APILEVEL);
-    PRINTF("Inside a library \n");
-    switch (args->command) {
-        case CHECK_ADDRESS:
-            // ensure result is zero if an exception is thrown
-            args->check_address->result = 0;
-            args->check_address->result =
-                handle_check_address(args->check_address, args->coin_config);
-            break;
-        case SIGN_TRANSACTION:
-            if (copy_transaction_parameters(args->create_transaction)) {
-                // never returns
-                handle_swap_sign_transaction(args->coin_config);
-            }
-            break;
-        case GET_PRINTABLE_AMOUNT:
-            // ensure result is zero if an exception is thrown (compatibility breaking, disabled
-            // until LL is ready)
-            // args->get_printable_amount->result = 0;
-            // args->get_printable_amount->result =
-            handle_get_printable_amount(args->get_printable_amount, args->coin_config);
-            break;
-        default:
-            break;
-    }
-}
-
-
-void init_coin_config(btchip_altcoin_config_t *coin_config);
-
-void library_main(struct libargs_s *args) {
-    btchip_altcoin_config_t coin_config;
-    if (args->coin_config == NULL) {
-        init_coin_config(&coin_config);
-        args->coin_config = &coin_config;
-    }
-    bool end = false;
-    /* This loop ensures that library_main_helper and os_lib_end are called
-     * within a try context, even if an exception is thrown */
-    while (1) {
-        BEGIN_TRY {
-            TRY {
-                if (!end) {
-                    library_main_helper(args);
-                }
-                os_lib_end();
-            }
-            FINALLY {
-                end = true;
-            }
-        }
-        END_TRY;
-    }
-}
