+++ src/main.c
@@ -37,6 +37,9 @@
 #include "legacy/include/btchip_context.h"
 #include "swap/swap_lib_calls.h"
 #include "swap/swap_globals.h"
+#include "swap/handle_swap_sign_transaction.h"
+#include "swap/handle_get_printable_amount.h"
+#include "swap/handle_check_address.h"
 
 #ifndef DISABLE_LEGACY_SUPPORT
 #include "legacy/main_old.h"
@@ -219,6 +222,14 @@ void app_main() {
             }
         } else {
 #endif
+            // if not Bitcoin or Bitcoin-testnet, we only support the legacy APDUS.
+            // to be removed once the apps are split
+            if (G_coin_config->bip32_pubkey_version != 0x0488B21E &&
+                G_coin_config->bip32_pubkey_version != 0x043587CF) {
+                io_send_sw(SW_CLA_NOT_SUPPORTED);
+                return;
+            }
+
             if (G_app_mode != APP_MODE_NEW) {
                 explicit_bzero(&G_command_state, sizeof(G_command_state));
 
@@ -282,6 +293,8 @@ void coin_main(btchip_altcoin_config_t *coin_config) {
 
     _Static_assert(sizeof(cx_sha256_t) <= 108, "cx_sha256_t too large");
     _Static_assert(sizeof(policy_map_key_info_t) <= 148, "policy_map_key_info_t too large");
+
+    G_app_mode = APP_MODE_UNINITIALIZED;
 
     btchip_altcoin_config_t config;
     if (coin_config == NULL) {
@@ -347,9 +360,62 @@ void coin_main(btchip_altcoin_config_t *coin_config) {
     app_exit();
 }
 
-__attribute__((section(".boot"))) int main(int arg0) {
-    G_app_mode = APP_MODE_UNINITIALIZED;
+static void library_main_helper(struct libargs_s *args) {
+    check_api_level(CX_COMPAT_APILEVEL);
+    PRINTF("Inside a library \n");
+    switch (args->command) {
+        case CHECK_ADDRESS:
+            // ensure result is zero if an exception is thrown
+            args->check_address->result = 0;
+            args->check_address->result =
+                handle_check_address(args->check_address, args->coin_config);
+            break;
+        case SIGN_TRANSACTION:
+            if (copy_transaction_parameters(args->create_transaction)) {
+                // never returns
+                handle_swap_sign_transaction(args->coin_config);
+            }
+            break;
+        case GET_PRINTABLE_AMOUNT:
+            // ensure result is zero if an exception is thrown (compatibility breaking, disabled
+            // until LL is ready)
+            // args->get_printable_amount->result = 0;
+            // args->get_printable_amount->result =
+            handle_get_printable_amount(args->get_printable_amount, args->coin_config);
+            break;
+        default:
+            break;
+    }
+}
 
+void init_coin_config(btchip_altcoin_config_t *coin_config);
+
+void library_main(struct libargs_s *args) {
+    btchip_altcoin_config_t coin_config;
+    if (args->coin_config == NULL) {
+        init_coin_config(&coin_config);
+        args->coin_config = &coin_config;
+    }
+    bool end = false;
+    /* This loop ensures that library_main_helper and os_lib_end are called
+     * within a try context, even if an exception is thrown */
+    while (1) {
+        BEGIN_TRY {
+            TRY {
+                if (!end) {
+                    library_main_helper(args);
+                }
+                os_lib_end();
+            }
+            FINALLY {
+                end = true;
+            }
+        }
+        END_TRY;
+    }
+}
+
+__attribute__((section(".boot"))) int main(int arg0) {
 #ifdef USE_LIB_BITCOIN
     BEGIN_TRY {
         TRY {
@@ -414,12 +480,8 @@ __attribute__((section(".boot"))) int main(int arg0) {
                 coin_main(args->coin_config);
             break;
         default:
-#ifndef DISABLE_LEGACY_SUPPORT
             // called as bitcoin or altcoin library
             library_main(args);
-#else
-            app_exit();
-#endif
     }
 #endif  // USE_LIB_BITCOIN
     return 0;
